Unified feature/experiment onboarding.
TODO:
- [x] New dynamic homepage for onboarding
- [ ] Add features to dashboard
- [ ] Hover effect in initial onboarding modal
