+++ packages/front-end/pages/features/index.tsx
@@ -11,18 +11,13 @@ import ValueDisplay from "../../components/Features/ValueDisplay";
 import { useRouter } from "next/router";
 import { useContext } from "react";
 import { UserContext } from "../../components/ProtectedPage";
-import EditAttributesModal from "../../components/Features/EditAttributesModal";
-import CodeSnippetModal from "../../components/Features/CodeSnippetModal";
 import track from "../../services/track";
-import GetStartedStep from "../../components/HomePage/GetStartedStep";
-import DocumentationLinksSidebar from "../../components/HomePage/DocumentationLinksSidebar";
+import FeaturesGetStarted from "../../components/HomePage/FeaturesGetStarted";
 
 export default function FeaturesPage() {
   const { project } = useDefinitions();
 
   const [modalOpen, setModalOpen] = useState(false);
-  const [attributeModalOpen, setAttributeModalOpen] = useState(false);
-  const [codeModalOpen, setCodeModalOpen] = useState(false);
   const router = useRouter();
   const { data, error, mutate } = useApi<{
     features: FeatureInterface[];
@@ -47,15 +42,6 @@ export default function FeaturesPage() {
     return <LoadingOverlay />;
   }
 
-  let step = -1;
-  if (!settings?.attributeSchema?.length) {
-    step = 0;
-  } else if (!settings?.sdkInstructionsViewed) {
-    step = 1;
-  } else if (!data?.features?.length) {
-    step = 2;
-  }
-
   return (
     <div className="contents container pagecontents">
       {modalOpen && (
@@ -68,12 +54,6 @@ export default function FeaturesPage() {
             });
           }}
         />
-      )}
-      {attributeModalOpen && (
-        <EditAttributesModal close={() => setAttributeModalOpen(false)} />
-      )}
-      {codeModalOpen && (
-        <CodeSnippetModal close={() => setCodeModalOpen(false)} />
       )}
       <div className="row mb-3">
         <div className="col">
@@ -122,71 +102,7 @@ export default function FeaturesPage() {
               </a>
             )}
           </h4>
-          <div className="row getstarted mb-3">
-            <div className="col-12 col-lg-8 ">
-              <div className={`card gsbox`} style={{ overflow: "hidden" }}>
-                <GetStartedStep
-                  current={step === 0}
-                  finished={settings?.attributeSchema?.length > 0}
-                  image="/images/attributes-icon.svg"
-                  title="1. Choose targeting attributes"
-                  text="Pick which user properties you want to pass into our SDKs. This enables you to use complex targeting rules and run experiments with your features."
-                  cta="Choose attributes"
-                  finishedCTA="Edit attributes"
-                  imageLeft={true}
-                  onClick={(finished) => {
-                    setAttributeModalOpen(true);
-                    if (!finished) {
-                      track("Viewed Attributes Modal", {
-                        source: "feature-onboarding",
-                      });
-                    }
-                  }}
-                />
-                <GetStartedStep
-                  current={step === 1}
-                  finished={settings?.sdkInstructionsViewed}
-                  className="border-top"
-                  image="/images/coding-icon.svg"
-                  title="2. Install our SDK"
-                  text="Integrate GrowthBook into your Javascript or React application. More languages and frameworks coming soon!"
-                  cta="View instructions"
-                  finishedCTA="View instructions"
-                  imageLeft={false}
-                  onClick={(finished) => {
-                    setCodeModalOpen(true);
-                    if (!finished) {
-                      track("Viewed Feature Integration Modal", {
-                        source: "feature-onboarding",
-                      });
-                    }
-                  }}
-                />
-                <GetStartedStep
-                  current={step === 2}
-                  finished={data?.features?.length > 0}
-                  className="border-top"
-                  image="/images/feature-icon.svg"
-                  title="3. Add your first feature"
-                  text="Create a feature within GrowthBook. It could be a simple ON/OFF flag or a configurable property like a color or copy for a headline."
-                  cta="Add first feature"
-                  finishedCTA="Add a feature"
-                  imageLeft={true}
-                  onClick={(finished) => {
-                    setModalOpen(true);
-                    if (!finished) {
-                      track("Viewed Feature Modal", {
-                        source: "feature-onboarding",
-                      });
-                    }
-                  }}
-                />
-              </div>
-            </div>
-            <div className="d-none d-lg-block col-lg-4">
-              <DocumentationLinksSidebar />
-            </div>
-          </div>
+          <FeaturesGetStarted features={data.features || []} mutate={mutate} />
           {!stepsRequired && <h4 mt-3>All Features</h4>}
         </div>
       ) : (
