+++ packages/front-end/pages/getstarted.tsx
@@ -1,32 +1,55 @@
 import { ExperimentInterfaceStringDates } from "back-end/types/experiment";
 import React from "react";
+import { FeatureInterface } from "../../back-end/types/feature";
 import GetStarted from "../components/HomePage/GetStarted";
 import LoadingOverlay from "../components/LoadingOverlay";
 import useApi from "../hooks/useApi";
 import { useDefinitions } from "../services/DefinitionsContext";
 
 const getStartedPage = (): React.ReactElement => {
-  const { data, mutate, error } = useApi<{
-    experiments: ExperimentInterfaceStringDates[];
-  }>("/experiments");
   const { ready, error: definitionsError } = useDefinitions();
 
-  if (error || definitionsError) {
+  const {
+    data: experiments,
+    error: experimentsError,
+    mutate: mutateExperiments,
+  } = useApi<{
+    experiments: ExperimentInterfaceStringDates[];
+  }>(`/experiments`);
+
+  const {
+    data: features,
+    error: featuresError,
+    mutate: mutateFeatures,
+  } = useApi<{
+    features: FeatureInterface[];
+  }>(`/feature`);
+
+  if (featuresError || experimentsError || definitionsError) {
     return (
       <div className="alert alert-danger">
-        An error occurred: {error?.message || definitionsError}
+        An error occurred:{" "}
+        {featuresError?.message ||
+          experimentsError?.message ||
+          definitionsError}
       </div>
     );
   }
 
-  if (!data || !ready) {
+  if (!experiments || !features || !ready) {
     return <LoadingOverlay />;
   }
 
   return (
     <>
-      <div className="container-fluid pagecontents getstarted">
-        <GetStarted experiments={data.experiments} mutate={mutate} />
+      <div className="container pagecontents position-relative">
+        <GetStarted
+          experiments={experiments?.experiments || []}
+          features={features?.features || []}
+          mutateExperiments={mutateExperiments}
+          mutateFeatures={mutateFeatures}
+          onboardingType={null}
+        />
       </div>
     </>
   );
